+++ plone/protect/__init__.py
@@ -4,3 +4,4 @@ from plone.protect.authenticator import createToken
 from plone.protect.authenticator import CustomCheckAuthenticator
 from plone.protect.postonly import check as PostOnly
 from plone.protect.utils import protect
+from plone.protect import monkey